# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Champion Skills Academy API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface DeleteResponse {
  success: boolean;
  message: string;
}

export interface AdminLoginInput {
  password: string;
}

export interface AdminAuthResult {
  success: boolean;
  message: string;
}

export interface AdminMe {
  authenticated: boolean;
}

export type RegistrationStatsByGender = {
  male: number;
  female: number;
};

export interface RegistrationStats {
  total: number;
  thisMonth: number;
  thisWeek: number;
  byGender: RegistrationStatsByGender;
}

export interface ContactInput {
  name: string;
  email: string;
  phone?: string;
  message: string;
}

export type RegistrationInputGender = typeof RegistrationInputGender[keyof typeof RegistrationInputGender];


export const RegistrationInputGender = {
  male: 'male',
  female: 'female',
} as const;

export type RegistrationInputShirtSize = typeof RegistrationInputShirtSize[keyof typeof RegistrationInputShirtSize];


export const RegistrationInputShirtSize = {
  YS: 'YS',
  YM: 'YM',
  YL: 'YL',
  AS: 'AS',
  AM: 'AM',
  AL: 'AL',
  AXL: 'AXL',
} as const;

export interface RegistrationInput {
  playerFirstName: string;
  playerLastName: string;
  dateOfBirth: string;
  age: number;
  gender: RegistrationInputGender;
  address: string;
  city: string;
  state: string;
  zip: string;
  school: string;
  grade: string;
  shirtSize: RegistrationInputShirtSize;
  parentName: string;
  parentRelationship: string;
  parentPhone: string;
  parentEmail: string;
  hearAboutUs: string;
  emergencyContact1Name: string;
  emergencyContact1Relationship: string;
  emergencyContact1Phone: string;
  /** @nullable */
  emergencyContact2Name?: string | null;
  /** @nullable */
  emergencyContact2Relationship?: string | null;
  /** @nullable */
  emergencyContact2Phone?: string | null;
  hasMedicalConditions: boolean;
  /** @nullable */
  medicalConditionsExplain?: string | null;
  takesMedication: boolean;
  /** @nullable */
  medicationList?: string | null;
  /** @nullable */
  allergies?: string | null;
  /** @nullable */
  medicalNotes?: string | null;
  liabilityAgreed: boolean;
  liabilitySignature: string;
  liabilityDate: string;
  attendanceAgreed: boolean;
  attendanceSignature: string;
  attendanceDate: string;
  parentConductAgreed: boolean;
  parentConductSignature: string;
  parentConductDate: string;
  playerConductAgreed: boolean;
  playerConductSignature: string;
  playerConductDate: string;
  photoReleaseGranted: boolean;
  photoReleaseParentName: string;
  photoReleaseSignature: string;
  photoReleaseDate: string;
}

export interface Registration {
  id: number;
  playerFirstName: string;
  playerLastName: string;
  dateOfBirth: string;
  age: number;
  gender: string;
  address: string;
  city: string;
  state: string;
  zip: string;
  school: string;
  grade: string;
  shirtSize: string;
  parentName: string;
  parentRelationship: string;
  parentPhone: string;
  parentEmail: string;
  hearAboutUs: string;
  emergencyContact1Name: string;
  emergencyContact1Relationship: string;
  emergencyContact1Phone: string;
  /** @nullable */
  emergencyContact2Name?: string | null;
  /** @nullable */
  emergencyContact2Relationship?: string | null;
  /** @nullable */
  emergencyContact2Phone?: string | null;
  hasMedicalConditions: boolean;
  /** @nullable */
  medicalConditionsExplain?: string | null;
  takesMedication: boolean;
  /** @nullable */
  medicationList?: string | null;
  /** @nullable */
  allergies?: string | null;
  /** @nullable */
  medicalNotes?: string | null;
  liabilityAgreed: boolean;
  liabilitySignature: string;
  liabilityDate: string;
  attendanceAgreed: boolean;
  attendanceSignature: string;
  attendanceDate: string;
  parentConductAgreed: boolean;
  parentConductSignature: string;
  parentConductDate: string;
  playerConductAgreed: boolean;
  playerConductSignature: string;
  playerConductDate: string;
  photoReleaseGranted: boolean;
  photoReleaseParentName: string;
  photoReleaseSignature: string;
  photoReleaseDate: string;
  createdAt: string;
}

export type ListRegistrationsParams = {
search?: string;
dateFrom?: string;
dateTo?: string;
};