# lib/api-zod/src/generated/types/deleteResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Champion Skills Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface DeleteResponse {
  success: boolean;
  message: string;
}