# lib/api-zod/src/generated/types/registration.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Champion Skills Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface Registration {
  id: number;
  playerFirstName: string;
  playerLastName: string;
  dateOfBirth: string;
  age: number;
  gender: string;
  address: string;
  city: string;
  state: string;
  zip: string;
  school: string;
  grade: string;
  shirtSize: string;
  parentName: string;
  parentRelationship: string;
  parentPhone: string;
  parentEmail: string;
  hearAboutUs: string;
  emergencyContact1Name: string;
  emergencyContact1Relationship: string;
  emergencyContact1Phone: string;
  /** @nullable */
  emergencyContact2Name?: string | null;
  /** @nullable */
  emergencyContact2Relationship?: string | null;
  /** @nullable */
  emergencyContact2Phone?: string | null;
  hasMedicalConditions: boolean;
  /** @nullable */
  medicalConditionsExplain?: string | null;
  takesMedication: boolean;
  /** @nullable */
  medicationList?: string | null;
  /** @nullable */
  allergies?: string | null;
  /** @nullable */
  medicalNotes?: string | null;
  liabilityAgreed: boolean;
  liabilitySignature: string;
  liabilityDate: string;
  attendanceAgreed: boolean;
  attendanceSignature: string;
  attendanceDate: string;
  parentConductAgreed: boolean;
  parentConductSignature: string;
  parentConductDate: string;
  playerConductAgreed: boolean;
  playerConductSignature: string;
  playerConductDate: string;
  photoReleaseGranted: boolean;
  photoReleaseParentName: string;
  photoReleaseSignature: string;
  photoReleaseDate: string;
  createdAt: string;
}