# lib/api-zod/src/generated/types/registrationStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Champion Skills Academy API
 * OpenAPI spec version: 0.1.0
 */
import type { RegistrationStatsByGender } from './registrationStatsByGender';

export interface RegistrationStats {
  total: number;
  thisMonth: number;
  thisWeek: number;
  byGender: RegistrationStatsByGender;
}